requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
fastapi==0.112.2
ffmpeg==1.4
ffpyplayer==4.5.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
iniconfig==1.1.1
intel-cmplr-lib-ur==2025.0.2
intel-openmp==2025.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mkl==2025.0.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.1
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2
numba==0.60.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.0.0
proglog==0.1.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
PySocks==1.7.1
pytest==7.4.4
pytest-flask==1.3.0
python-dateutil==2.9.0
pytube==15.0.0
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.0
starlette==0.38.2
sympy==1.13.1
tbb==2022.0.0
tcmlib==1.2.0
tiktoken==0.7.0
torch==2.5.1
tqdm==4.66.5
typing_extensions==4.11.0
tzdata==2024.2
umf==0.9.1
urllib3==2.2.2
uvicorn==0.20.0
Werkzeug==3.0.3
zipp==3.20.0
zstandard==0.23.0
